+++ cmake/nfxGraphicsSources.cmake
@@ -6,7 +6,9 @@ set(NFX_GRAPHICS_SOURCES
     src/gl/material/Material.cpp
     src/gl/mesh/Mesh.cpp
     src/gl/mesh/Primitive.cpp
+    src/gl/pipeline/passes/AxesPass.cpp
     src/gl/pipeline/passes/GeometryPass.cpp
+    src/gl/pipeline/passes/GridPass.cpp
     src/gl/pipeline/passes/PresentPass.cpp
     src/gl/pipeline/Renderer.cpp
     src/gl/pipeline/RenderState.cpp