+++ include/nfx/Graphics.h
@@ -22,7 +22,9 @@
 // Pipeline
 #include "nfx/graphics/gl/pipeline/frame/FrameData.h"
 #include "nfx/graphics/gl/pipeline/frame/RenderResources.h"
+#include "nfx/graphics/gl/pipeline/passes/AxesPass.h"
 #include "nfx/graphics/gl/pipeline/passes/GeometryPass.h"
+#include "nfx/graphics/gl/pipeline/passes/GridPass.h"
 #include "nfx/graphics/gl/pipeline/passes/PresentPass.h"
 #include "nfx/graphics/gl/pipeline/passes/RenderPass.h"
 #include "nfx/graphics/gl/pipeline/queue/RenderCommand.h"